sizes should be int on sender/receiver

+++ src/ck-core/charm.h
@@ -131,7 +131,7 @@ extern void registerChareMsgRecvExtCallback(void (*cb)(int, void*, int, int, cha
 extern void registerGroupMsgRecvExtCallback(void (*cb)(int, int, int, char *, int));
 extern void registerArrayMsgRecvExtCallback(void (*cb)(int, int, int *, int, int, char *, int));
 #if CMK_CUDA
-extern void registerArrayMsgGPUDirectRecvExtCallback(void (*cb)(int, int, int*, int, int, long*, void *, int, char*,int));
+extern void registerArrayMsgGPUDirectRecvExtCallback(void (*cb)(int, int, int*, int, int, int*, void *, int, char*,int));
 #endif
 extern void registerArrayBcastRecvExtCallback(void (*cb)(int, int, int, int, int*, int, int, char *, int));
 extern void registerArrayElemLeaveExtCallback(int (*cb)(int, int, int *, char**, int));
@@ -501,7 +501,7 @@ extern void CkChareExtSendWithDeviceData(int aid, int *idx, int ndims,
                                          int epIdx, int num_bufs, char **bufs,
                                          int *buf_sizes,
                                          long *devBufPtrs,
-                                         long *devBufSizesInBytes,
+                                         int *devBufSizesInBytes,
                                          long *streamPtrs, int numDevBufs
                                         );
 extern int CkDeviceBufferSizeInBytes();
